Capturing and storing the exam information to the database
this push reworks the exam landing page so it works with our springboot
back-end, and made it so the user cannot enter a negative number or a
zero. It also sets up methods for capturing the data from exam landing
page and routing the user to their selected option in the instructor
controller. It also adds fields to the exam.java for end time of the
exam.