Association of Students/Instructors. Deleting students with association
This push adds functionality behind the pages that associate courses
with students and instructors with their courses.

It also fixes a bug where if a student was associated with a class, and
that student were to be deleted, it would then delete all the students
that were associated with that course. This no longer happens.