Adding the studentID to the end of their password during roster upload
This push makes it so when the class roster is being uploaded, the
students that don't already exist are created with a password that has
their ID tacked to the end of it adding additional security to accounts
that have not been accessed for the first time. It also changes the
method of checking if the student and instructor already have an
account. Rather then student and instructor ID's it checks the username
to make sure they were not already created. This is a better method
because when logging in the user uses their username not their ID and if
the username already exists then the student can't login.