student update bugs, instructor can change student password.
I added student password change for instructors, and fixed it under both
schedule manager and instructor so its able to update students without
removing associations from the student. Before, when you made an edit to
the student is basically just overwrote the old student with a new one.
I updated it so it is simply updating the existing student entity.